Removed italics from integrals
The standard integral characters are at an odd angle, rotated counter-clockwise to where they should be. I addressed this by italicizing the font for integrals, but the appearance of italic font varies depending on the system. It looked right on Windows, but wasn't corrected enough on Linux and WASM. This fix rotates the integral characters instead of italicizing the font.